Debugging and moving some sections to core
