+++ src/auxiliary.py
@@ -5,12 +5,12 @@ from math import sqrt, ceil
 from typing import Sequence
 
 
+
+# ------------ Core Section ------------
+
 # Constants needed for down scaling spectra and images
 fwhm_factor = np.sqrt(8*np.log(2))
 hanning_factor = 1129/977
-
-
-# Math operations
 
 def get_resolution(array: Sequence):
     return np.mean(np.diff(array)) * hanning_factor
@@ -28,7 +28,7 @@ def is_smooth(array: Sequence|np.ndarray):
     diff2 = np.diff(np.diff(array, axis=0), axis=0)
     return np.all(diff2 <= 0) | np.all(diff2 >= 0)
 
-def integrate(array: Sequence|np.ndarray, step: int|float, precisely=False):
+def integrate(array: Sequence|np.ndarray, step: int|float, precisely: bool = False):
     """
     Integration along the spectral axis.
     Uses the rectangle method by default and Riemann sum with midpoint in the "precise" mode.
@@ -161,13 +161,13 @@ def interpolating(x0: Sequence, y0: np.ndarray, x1: Sequence, step: int|float) -
         y0 = custom_interp(y0, k=11+i)
     return linear_interp(x0, y0, x1)
 
-def higher_dim(scope: Sequence, times: int, axis: int = 1):
+def higher_dim(arr: Sequence|int|float, times: int, axis: int = 1):
     """ Gets the array and repeats it along a new dimension """
-    return np.repeat(np.expand_dims(scope, axis=axis), times, axis=axis)
+    return np.repeat(np.expand_dims(arr, axis=axis), times, axis=axis)
 
-def scope2cube(scope: Sequence, shape: tuple[int, int]):
+def scope2cube(arr: Sequence, shape: tuple[int, int]):
     """ Gets the 1D array and expands its dimensions to a 3D array based on the 2D slice shape """
-    return np.repeat(np.repeat(np.expand_dims(scope, axis=(1, 2)), shape[0], axis=1), shape[1], axis=2)
+    return np.repeat(np.repeat(np.expand_dims(arr, axis=(1, 2)), shape[0], axis=1), shape[1], axis=2)
 
 def expand_1D_array(arr: np.ndarray, shape: int|tuple):
     """ Gets the 1D array and expands its dimensions to a 2D or 3D array based on the slice shape """
@@ -246,6 +246,187 @@ def extrapolating(x: np.ndarray, y: np.ndarray, scope: np.ndarray, step: int|flo
             y = np.append(y, y1, axis=0)
     return x, y
 
+# Blackbody spectra
+
+h = 6.626e-34 # Planck constant
+c = 299792458 # Speed of light
+k = 1.381e-23 # Boltzmann constant
+const1 = 2 * np.pi * h * c * c
+const2 = h * c / k
+r = 6.957e8 # Solar radius, meters
+au = 149597870700 # astronomical unit, meters
+w = (1 - np.sqrt(1 - (r / au)**2)) / 2 # dilution to compare with Solar light on Earth
+temp_coef_to_make_it_work = 1.8 / 0.448 # 1.8 is expected of irradiance(500, 5770), 0.448 is actual. TODO
+
+# Now it does not work as intended: the spectrum should be comparable to the sun_SI
+def irradiance(nm: int|float|np.ndarray, T: int|float) -> float|np.ndarray:
+    m = nm / 1e9
+    return temp_coef_to_make_it_work * w * const1 / (m**5 * (np.exp(const2 / (m * T)) - 1)) / 1e9 # per m -> per nm
+
+
+
+# ------------ Database Processing Section ------------
+
+def parse_value_sd(data: float|Sequence[float]):
+    """ Guarantees the output of the value and its sd for variable input """
+    if isinstance(data, Sequence) and len(data) == 2:
+        value, sd = data
+    elif isinstance(data, (int, float)):
+        value = data
+        sd = None
+    else:
+        print(f'Invalid data input: {data}. Must be a numeric value or a [value, sd] list. Returning None.')
+        value = sd = None
+    return value, sd
+
+def mag2irradiance(mag: int|float|np.ndarray, zero_point: float = 1.):
+    """ Converts magnitudes to irradiance (by default in Vega units) """
+    return zero_point * 10**(-0.4 * mag)
+
+def sd_mag2sd_irradiance(sd_mag: int|float|np.ndarray, irradiance: int|float|np.ndarray):
+    """
+    Converts standard deviation of the magnitude to a irradiance standard deviation.
+
+    The formula is derived from the error propagation equation:
+    I(mag) = zero_point ∙ 10^(-0.4 mag)
+    sd_I² = (d I / d mag)² ∙ sd_mag²
+    I' = zero_point∙(10^(-0.4 mag))' = zero_point∙10^(-0.4 mag)∙ln(10^(-0.4)) = I∙(-0.4) ln(10)
+    sd_I = |I'| ∙ sd_mag = 0.4 ln(10) ∙ I ∙ sd_mag
+    """
+    return 0.4 * np.log(10) * irradiance * sd_mag
+
+def color_index_splitter(index: str):
+    """
+    Dashes in filter names are allowed in the SVO Filter Profile Service.
+    This function should fix all or most of the problems caused.
+    """
+    try:
+        filter1, filter2 = index.split('-')
+    except ValueError:
+        dotpart1, dotpart2, dotpart3 = index.split('.') # one dot per full filter name
+        dashpart1, dashpart2 = dotpart2.split('-', 1)
+        filter1 = f'{dotpart1}.{dashpart1}'
+        filter2 = f'{dashpart2}.{dotpart3}'
+    return filter1, filter2
+
+def color_indices_parser(indices: dict):
+    """
+    Converts color indices to linear brightness, assuming mag=0 in the first filter.
+    Each new color index must refer to a previously specified one.
+    Note: The output order may sometimes not be in ascending wavelength order.
+    This can be corrected by knowing the filter profiles, which better to do outside the function.
+
+    For standard deviations the error propagation equation is used:
+    f(x, y) = x - y
+    sd_f² = (df/dx)² sd_x² + (df/dy)² sd_y² = sd_x² + sd_y²
+    where x, y are magnitudes and f is a color index.
+
+    Finding standard deviations of a photospectrum built from color indices is an ill-posed problem:
+    it's just like with integrating, we loose constant after differentiation (color indices is
+    a discrete differential form of a photospectrum).
+    For the photospectrum itself, it's not a problem: the solutions dimension is just scaling
+    the spectrum on a constant (it's pretty obvious that color indices always lost brightness
+    information).
+    But the solutions space for their standard deviations is more complex, I found its geometric
+    interpretation: since the standard deviations subtracting rule is, in fact, the Pythagorean theorem,
+    the solutions space is the same as if you try to build a line of right triangles, for each one
+    the next cathetus is linked to a previous cathetus by their square.
+    N hypotenuses are standard deviations of color indices, and N+1 different cathetes are the sought
+    standard deviations of the photospectrum.
+    The whole triangle line possible positions can be described by just one parameter (1D parametric
+    space of solutions).
+    For simplicity, I will choose the first cathetus (the first sought standard deviation) as
+    a variable of this space.
+    Such triangles can "collapse" if the previous cathetus is greater than the next hypotenuse!
+    I tried to find an analytical solution, but the requirement for the optimal solution I derived
+    suggested that about a half of the triangles should be collapsed.
+
+    In the numerical approach I use, some solutions are collapsed (the `try-except` code block),
+    but there are a some range of possible solutions too, which one to choose?
+    I decided to choose the solution with minimal standard deviation of standard deviations it gives.
+    To tighten the solution selection criteria, it can be assumed that the size of the standard deviation
+    is inversely proportional to the root of the irradiance (in the Poisson noise approximation).
+    So it's better to minimize the differences not between the stds of magnitudes, but between
+    the stds of scaled irradiances.
+    """
+    first_color_index = tuple(indices.keys())[0]
+    filter0, _ = color_index_splitter(first_color_index)
+    _, sd0 = parse_value_sd(indices[first_color_index])
+    # Just photospectrum calculation
+    uncertainty_flag = False
+    filters = {filter0: 0} # mag=0 for the first point (arbitrarily)
+    for key, value in indices.items():
+        bluer_filter, redder_filter = color_index_splitter(key)
+        mag, sd = parse_value_sd(value)
+        if sd is not None:
+            uncertainty_flag = True
+        if bluer_filter in filters:
+            filters |= {redder_filter: filters[bluer_filter] - mag}
+        else:
+            filters |= {bluer_filter: filters[redder_filter] + mag}
+    irradiance = mag2irradiance(np.array(tuple(filters.values())))
+    sd = np.zeros_like(irradiance)
+    # Uncertainty calculation
+    if uncertainty_flag:
+        shot_noise_factor = np.sqrt(irradiance) # common Poisson noise factor
+        old_sd_of_sd = np.inf
+        for sd_assumed in np.linspace(0, sd0, 1001):
+            impossible_assumption = False
+            # Numerically select the best value of the standard deviation of the first point,
+            # on which all other standard deviations clearly depend
+            filters = {filter0: sd_assumed}
+            for key, value in indices.items():
+                bluer_filter, redder_filter = color_index_splitter(key)
+                _, sd = parse_value_sd(value)
+                try:
+                    if bluer_filter in filters:
+                        filters |= {redder_filter: sqrt(sd**2 - filters[bluer_filter]**2)}
+                    else:
+                        filters |= {bluer_filter: sqrt(sd**2 - filters[redder_filter]**2)}
+                except ValueError:
+                    # This means that the difference under the root is negative
+                    # and the initial standard deviation assumption is not possible
+                    impossible_assumption = True
+                    break
+            if not impossible_assumption:
+                new_sd = sd_mag2sd_irradiance(np.array(tuple(filters.values())), irradiance)
+                # Finding the minimum deviation between sd as solution quality criterion
+                # The standard deviations are scaled by the Poisson noise factor
+                new_sd_of_sd = np.std(new_sd * shot_noise_factor)
+                if new_sd_of_sd < old_sd_of_sd:
+                    old_sd = new_sd
+                    old_sd_of_sd = new_sd_of_sd
+                    continue
+                else:
+                    # Means that the best values of standard deviations were found
+                    # in the last iteration and they started to diverge
+                    sd = old_sd
+                    break
+    return filters.keys(), irradiance, sd
+
+def phase_function2phase_integral(name: str, params: dict):
+    """ Determines phase integral from the phase function """
+    phase_integral = phase_integral_sd = None
+    match name:
+        case 'HG':
+            g, g_sd = parse_value_sd(params['G'])
+            phase_integral = 0.290 + 0.684 * g
+            if g_sd is not None:
+                phase_integral_sd = 0.827 * g_sd # 0.827 ≈ sqrt(0.684)
+        case 'HG1G2':
+            g1, g1_sd = parse_value_sd(params['G_1'])
+            g2, g2_sd = parse_value_sd(params['G_2'])
+            phase_integral = 0.009082 + 0.4061 * g1 + 0.8092 * g2
+            if g1_sd is not None and g2_sd is not None:
+                phase_integral_sd = np.sqrt(0.4061 * g1_sd + 0.8092 * g2_sd)
+        case _:
+            print(f'Phase function with name {name} is not supported.')
+    return phase_integral, phase_integral_sd
+
+
+
+# ------------ Color Processing Section ------------
+
 def gamma_correction(arr0: np.ndarray):
     """ Applies gamma correction in CIE sRGB implementation to the array """
     arr1 = np.copy(arr0)
@@ -265,6 +446,9 @@ def export_colors(rgb: tuple):
             mx = l
     w = 8 if mx < 8 else mx+1
     return ''.join([i.ljust(w) for i in lst])
+
+
+# ------------ Other ------------
 
 def get_flag_index(flags: tuple):
     """ Returns index of active radio button """
@@ -304,22 +488,3 @@ def normalize_string(string: str):
         if char not in illegal_chars:
             result += char
     return result
-
-
-
-# Blackbody spectra
-
-h = 6.626e-34 # Planck constant
-c = 299792458 # Speed of light
-k = 1.381e-23 # Boltzmann constant
-const1 = 2 * np.pi * h * c * c
-const2 = h * c / k
-r = 6.957e8 # Solar radius, meters
-au = 149597870700 # astronomical unit, meters
-w = (1 - np.sqrt(1 - (r / au)**2)) / 2 # dilution to compare with Solar light on Earth
-temp_coef_to_make_it_work = 1.8 / 0.448 # 1.8 is expected of irradiance(500, 5770), 0.448 is actual. TODO
-
-# Now it does not work as intended: the spectrum should be comparable to the sun_SI
-def irradiance(nm: int|float|np.ndarray, T: int|float) -> float|np.ndarray:
-    m = nm / 1e9
-    return temp_coef_to_make_it_work * w * const1 / (m**5 * (np.exp(const2 / (m * T)) - 1)) / 1e9 # per m -> per nm
