+++ src/core.py
@@ -1,8 +1,11 @@
 
 """
-Describes the main spectral data storage classes and related functions.
+Describes the main data storage classes and related functions.
 
-The hierarchy of core classes in a project:
+Naming:
+- ObjectName
+
+Data:
 - TrueColorToolsObject
 - - SpectralObject
 - - - 1D: Spectrum
@@ -11,6 +14,16 @@ The hierarchy of core classes in a project:
 - - PhotospectralObject
 - - - 1D: Photospectrum
 - - - 3D: PhotospectralCube
+
+Database:
+- NonReflectiveBody
+- ReflectiveBody
+
+Color:
+- ColorSystem
+- ColorImage
+- - ColorPoint
+
 
 (Self-complete classes system would include a 2D `PhotospectrumSet` and
 `SpectrumSet` instead of `FilterSystem`, with convolution operation defined.
@@ -29,8 +42,12 @@ import numpy as np
 from src.data_import import file_reader
 import src.auxiliary as aux
 import src.strings as tr
+import src.data_import as di
 import src.image_import as ii
 
+
+
+# ------------ Naming Section ------------
 
 class ObjectName:
     """
@@ -144,6 +161,8 @@ class ObjectName:
 
 
 
+# ------------ Data Section ------------
+
 # TCT was written in an effort to work not only with the optical range, but with any, depending on the data.
 # But too long and heterogeneous FITS files demanded to set the upper limit of the range to mid-wavelength infrared (3 μm).
 nm_red_limit = 3000 # nm
@@ -186,7 +205,7 @@ class _TrueColorToolsObject:
         # TODO: uncertainty processing
         # TODO: problems with cubes?
         output = deepcopy(self)
-        if isinstance(where, (str, int, float)):
+        if isinstance(where, str|int|float):
             where = get_filter(where)
         current_br = self @ where
         if current_br == 0:
@@ -213,7 +232,7 @@ class _TrueColorToolsObject:
         return output
     
     def __mul__(self, other: int|float|np.ndarray):
-        if isinstance(other, (int, float, np.ndarray)):
+        if isinstance(other, int|float|np.ndarray):
             return self.apply_linear_operator(mul, other)
         return NotImplemented
     
@@ -221,7 +240,7 @@ class _TrueColorToolsObject:
         return self.__mul__(other)
     
     def __truediv__(self, other: int|float|np.ndarray):
-        if isinstance(other, (int, float, np.ndarray)):
+        if isinstance(other, int|float|np.ndarray):
             return self.apply_linear_operator(truediv, other)
         return NotImplemented
 
@@ -394,6 +413,10 @@ class _SpectralObject(_TrueColorToolsObject):
             extrapolated.br = extrapolated.get_br_in_range(start, end)
         return extrapolated
     
+    def is_edges_zeroed(self) -> bool:
+        """ Checks that the first and last brightness entries on the spectral axis are zero """
+        return np.all(self.br[0] == 0) and np.all(self.br[-1] == 0)
+    
     def __matmul__(self, other):
         return other.__rmatmul__(self)
     
@@ -413,31 +436,37 @@ class _SpectralObject(_TrueColorToolsObject):
         - PhotospectralCube @ FilterSystem -> PhotospectralCube
         """
         # TODO: uncertainty processing
-        operand1 = other.to_scope(self.nm) # also reconstructs PhotospectralObj to SpectralObj
-        operand2 = self.to_scope(other.nm)
-        # Why Spectrum/FilterSystem is also extrapolated?
-        # For the cases of bolometric albedo operations such as `Sun @ Mercury`.
-        match operand1.ndim:
-            case 1: # Spectrum or Photospectrum
-                match operand2.ndim:
-                    case 1: # Spectrum
-                        return aux.integrate(operand1.br * operand2.br, nm_step)
-                    case 2: # FilterSystem
-                        br = aux.integrate((operand2.br.T * operand1.br).T, nm_step)
-                        return Photospectrum(operand2, br, name=operand1.name)
-            case 2: # FilterSystem
-                print(f'# Note for the {operand1.__class__.__name__} "{operand1.name}"')
-                print(f'- Convolution of a filter system makes no physical sense.')
-            case 3: # SpectralCube or PhotospectralCube
-                match operand2.ndim:
-                    case 1: # Spectrum
-                        return aux.integrate((operand1.br.T * operand2.br).T, nm_step)
-                    case 2: # FilterSystem
-                        br = np.empty((len(operand2), *operand1.shape))
-                        for i in range(len(operand2)):
-                            profile = operand2.br[:,i]
-                            br[i] = aux.integrate((operand1.br.T * profile).T, nm_step)
-                        return PhotospectralCube(operand2, br, name=operand1.name)
+        if self.is_edges_zeroed():
+            # No filter extrapolation required
+            operand1 = other.to_scope(self.nm, crop=True)
+            operand2 = self
+        else:
+            # Why Spectrum/FilterSystem needs extrapolation?
+            # For the cases of bolometric albedo operations such as `Sun @ Mercury`.
+            operand1 = other.to_scope(self.nm, crop=False)
+            operand2 = self.to_scope(other.nm, crop=False)
+        # other.to_scope() reconstructed PhotospectralObj to SpectralObj, so 4 options left:
+        if isinstance(operand1, Spectrum):
+            # Spectrum @ Spectrum
+            if isinstance(operand2, Spectrum):
+                return aux.integrate(operand1.br * operand2.br, nm_step)
+            # Spectrum @ FilterSystem
+            elif isinstance(operand2, FilterSystem):
+                br = aux.integrate((operand2.br.T * operand1.br).T, nm_step)
+                return Photospectrum(operand2, br, name=operand1.name)
+        elif isinstance(operand1, SpectralCube):
+            # SpectralCube @ Spectrum
+            if isinstance(operand2, Spectrum):
+                return aux.integrate((operand1.br.T * operand2.br).T, nm_step)
+            # SpectralCube @ FilterSystem
+            # A loop-less implementation would require a 4D array,
+            # which most computers do not have enough memory for.
+            elif isinstance(operand2, FilterSystem):
+                br = np.empty((len(operand2), *operand1.shape))
+                for i in range(len(operand2)):
+                    profile = operand2.br[:,i]
+                    br[i] = aux.integrate((operand1.br.T * profile).T, nm_step)
+                return PhotospectralCube(operand2, br, name=operand1.name)
         return NotImplemented
 
 
@@ -450,6 +479,7 @@ class Spectrum(_SpectralObject):
     - `br` (np.ndarray): array of "brightness" in energy density units (not a photon counter)
     - `sd` (np.ndarray): optional array of standard deviations
     - `name` (ObjectName): name as an instance of a class that stores its components
+    - `photospectrum` (Photospectrum): optional, way to store the pre-reconstructed data
     """
     
     def __init__(self, nm: Sequence, br: Sequence, sd: Sequence = None,
@@ -690,7 +720,7 @@ class FilterSystem(_SpectralObject):
         filters = list(filters)
         names = []
         for i, profile in enumerate(filters):
-            if isinstance(profile, (str, int, float)):
+            if isinstance(profile, str|int|float):
                 profile = get_filter(profile)
                 filters[i] = profile
             min_arr.append(profile.nm[0])
@@ -708,6 +738,7 @@ class FilterSystem(_SpectralObject):
         for i in range(len(self)):
             yield self[i]
     
+    @lru_cache(maxsize=32)
     def __getitem__(self, index: int) -> Spectrum:
         """ Returns the filter profile with extra zeros trimmed off """
         profile = self.br[:, index]
@@ -908,3 +939,407 @@ class PhotospectralCube(_PhotospectralObject, _Cube):
     def stub(name=None):
         """ Initializes an object in case of the data problems """
         return PhotospectralCube(FilterSystem.from_list(('Generic_Bessell.B', 'Generic_Bessell.V')), np.ones((2, 1, 1)), name=name)
+
+
+
+# ------------ Database Processing Section ------------
+
+sun_SI = Spectrum.from_file('spectra/files/CALSPEC/sun_reference_stis_002.fits', name='Sun') # W / (m² nm)
+sun_in_V = sun_SI @ get_filter('Generic_Bessell.V')
+sun_norm = sun_SI.scaled_at(get_filter('Generic_Bessell.V'))
+sun_filter = sun_SI.normalize()
+
+vega_SI = Spectrum.from_file('spectra/files/CALSPEC/alpha_lyr_stis_011.fits', name='Vega') # W / (m² nm)
+vega_in_V = vega_SI @ get_filter('Generic_Bessell.V')
+vega_norm = vega_SI.scaled_at(get_filter('Generic_Bessell.V'))
+
+
+class NonReflectiveBody:
+    """ High-level processing class, specializing on photometry of a physical body with not specified reflectance. """
+
+    def __init__(self, name: ObjectName, tags: Sequence, spectrum: Spectrum):
+        """
+        Args:
+        - `name` (ObjectName): name as an instance of a class that stores its components 
+        - `tags` (Sequence): list of categories that specify the physical body
+        - `spectrum` (Spectrum): not assumed to be scaled
+        """
+        self.name = name
+        self.tags = tags
+        self.spectrum = spectrum
+    
+    def get_spectrum(self, mode: str):
+        """
+        Returns the spectrum as the first argument, and the `estimated=False` bool status as the second one.
+        Albedo not determined for NonReflectiveBody, so it can't be "estimated", but we need output compatibility with ReflectiveBody.
+        """
+        if 'star' in self.tags:
+            return self.spectrum, False # means it's an emitter and we need to render it
+        else:
+            return Spectrum.stub(self.name), False # means we don't need to render it
+
+
+class ReflectiveBody:
+    """
+    High-level processing class, specializing on reflectance photometry of a physical body.
+
+    Albedo formatting rules:
+    1. `geometric_albedo` and `spherical_albedo` can be a boolean type or in [filter/nm, br, sd] format
+    2. both values by default `false`
+    3. no albedo is specified and "star" in tags → emitter, else black output
+    4. one albedo is specified → another one is estimated with the phase integral model
+
+    Phase integral model by Shevchenko et al.: https://ui.adsabs.harvard.edu/abs/2019A%26A...626A..87S/abstract
+    q = 0.359 (± 0.005) + 0.47 (± 0.03) p, where `p` is geometric albedo.
+    By definition, spherical albedo A is p∙q.
+    """
+
+    def __init__(
+            self, name: ObjectName, tags: Sequence,
+            geometric: Spectrum = None, spherical: Spectrum = None,
+            phase_integral: tuple[float, float] = None
+        ):
+        """
+        Args:
+        - `name` (ObjectName): name as an instance of a class that stores its components 
+        - `tags` (Sequence): list of categories that specify the physical body
+        - `geometric` (Spectrum): represents geometric albedo
+        - `spherical` (Spectrum): represents spherical albedo
+        - `phase_integral` (tuple[value, sd]): factor of transition from geometric albedo to spherical
+        """
+        self.name = name
+        self.tags = tags
+        self.spherical = spherical
+        self.geometric = geometric
+        self.phase_integral = phase_integral
+    
+    def get_spectrum(self, mode: str):
+        """
+        Returns the albedo-scaled spectrum as the first argument, and the `estimated` bool status as the second one.
+        `estimated = True` if the albedo was not known and estimated using a theoretical model.
+        """
+        match mode:
+            case 'geometric':
+                if self.geometric:
+                    return self.geometric, False
+                else:
+                    sphericalV = self.spherical @ get_filter('Generic_Bessell.V')
+                    if self.phase_integral is not None:
+                        phase_integral = self.phase_integral[0]
+                        geometricV = sphericalV / phase_integral
+                        estimated = False
+                    else:
+                        geometricV = (np.sqrt(0.359**2 + 4 * 0.47 * sphericalV) - 0.359) / (2 * 0.47)
+                        estimated = True
+                    return self.spherical.scaled_at(get_filter('Generic_Bessell.V'), geometricV), estimated
+            case 'spherical':
+                if self.spherical:
+                    return self.spherical, False
+                else:
+                    geometricV = self.geometric @ get_filter('Generic_Bessell.V')
+                    if self.phase_integral is not None:
+                        phase_integral = self.phase_integral[0]
+                        estimated = False
+                    else:
+                        phase_integral = 0.359 + 0.47 * geometricV
+                        estimated = True
+                    sphericalV = geometricV * phase_integral
+                    return self.geometric.scaled_at(get_filter('Generic_Bessell.V'), sphericalV), estimated
+
+def _create_TCT_object(
+        name: ObjectName, nm: Sequence[int|float], filters: Sequence[str], br: Sequence,
+        sd: Sequence = None, calib: str = None, sun: bool = False
+    ):
+    """
+    Decides whether we are dealing with photospectrum or continuous spectrum
+    and calibrates the spectral object.
+    """
+    if len(nm) > 0:
+        TCT_obj = Spectrum.from_array(nm, br, sd, name=name)
+    elif len(filters) > 0:
+        TCT_obj = Photospectrum(FilterSystem.from_list(filters), br, sd, name=name)
+    else:
+        print(f'# Note for the database object "{name}"')
+        print(f'- No wavelength data. Spectrum stub object was created.')
+        TCT_obj = Spectrum.stub(name)
+    match calib:
+        case 'vega':
+            TCT_obj *= vega_norm
+        case 'ab':
+            TCT_obj = TCT_obj.convert_from_frequency_spectral_density()
+        case _:
+            pass
+    if sun:
+        TCT_obj /= sun_norm
+    return TCT_obj
+
+def database_parser(name: ObjectName, content: dict) -> NonReflectiveBody | ReflectiveBody:
+    """
+    Depending on the contents of the object read from the database, returns a class that has `get_spectrum()` method
+
+    Supported input keys of a database unit:
+    - `nm` (list): list of wavelengths in nanometers
+    - `br` (list): same-size list of "brightness" in energy density units (not a photon counter)
+    - `mag` (list): same-size list of magnitudes
+    - `sd` (list/number): same-size list of standard deviations or a general value
+    - `nm_range` (dict): `start`, `stop`, `step` keys defining a wavelength range
+    - `slope` (dict): `start`, `stop`, `power` keys defining a spectrum from spectrophotometric gradient
+    - `file` (str): path to a text or FITS file, recommended placing in `spectra` or `spectra_extras` folder
+    - `filters` (list): list of filter names (see `filters` folder), can be mixed with nm values if needed
+    - `color_indices` (list): dictionary of color indices, formatted `{'filter1-filter2': [br, (sd)]], …}`
+    - `photometric_system` (str): a way to bracket the name of the photometric system
+    - `calibration_system` (str): `Vega` or `AB` filters zero points calibration, `ST` is assumed by default
+    - `albedo` (bool/list): indicates data as albedo scaled or tells how to do it with `[filter/nm, [br, (sd)]]`
+    - `geometric_albedo` (bool/list): indicator of geometric/normal albedo data or how to scale to it
+    - `spherical_albedo` (bool/list): indicator of spherical albedo data or how to scale to it
+    - `bond_albedo` (number): sets spherical albedo scale using known solar spectrum
+    - `phase_integral` (number/list): factor of transition from geometric albedo to spherical (sd is optional)
+    - `phase_function` (list): function name and its parameters to compute phase integral (sd is optional)
+    - `br_geometric`, `br_spherical` (list): specifying unique spectra for different albedos
+    - `sd_geometric`, `sd_spherical` (list/number): corresponding standard deviations or a general value
+    - `sun_is_emitter` (bool): `true` to remove the reflected solar spectrum
+    - `tags` (list): strings categorizing the spectrum
+    """
+    br = []
+    sd = None
+    nm = [] # Spectrum object indicator
+    filters = [] # Photospectrum object indicator
+    if 'file' in content:
+        try:
+            nm, br, sd = di.file_reader(content['file'])
+        except Exception:
+            stub = Spectrum.stub()
+            nm = stub.nm
+            br = stub.br
+            print(f'# Note for the Spectrum object "{name}"')
+            print(f'- Something unexpected happened during external file reading. The data was replaced by a stub.')
+            print(f'- More precisely, {format_exc(limit=0).strip()}')
+    else:
+        # Brightness reading
+        if 'br' in content:
+            br = content['br']
+            if 'sd' in content:
+                sd = aux.higher_dim(content['sd'], len(br), axis=0)
+        elif 'mag' in content:
+            br = aux.mag2irradiance(np.array(content['mag']))
+            if 'sd' in content:
+                sd = aux.higher_dim(content['sd'], len(br), axis=0)
+                sd = aux.sd_mag2sd_irradiance(sd, br)
+        # Spectrum reading
+        if 'nm' in content:
+            nm = content['nm']
+        elif 'nm_range' in content:
+            nm_range = content['nm_range']
+            nm = np.arange(nm_range['start'], nm_range['stop']+1, nm_range['step'])
+            # important not to use aux.grid() here
+        elif 'slope' in content:
+            slope = content['slope']
+            nm = aux.grid(slope['start'], slope['stop'], nm_step)
+            br = (nm / nm[0])**slope['power']
+        # Photospectrum reading
+        elif 'filters' in content:
+            filters = content['filters']
+        elif 'color_indices' in content:
+            filters, br, sd = aux.color_indices_parser(content['color_indices'])
+        if 'photometric_system' in content:
+            # regular filter if name is string, else "delta-filter" (wavelength)
+            filters = [f'{content["photometric_system"]}.{short_name}' if isinstance(short_name, str) else short_name for short_name in filters]
+    calib = content['calibration_system'].lower() if 'calibration_system' in content else None
+    sun = 'sun_is_emitter' in content and content['sun_is_emitter']
+    geometric = spherical = None
+    if len(br) == 0:
+        if 'br_geometric' in content:
+            br_geom = content['br_geometric']
+            sd_geom = aux.higher_dim(content['sd_geometric'], len(br_geom), axis=0) if 'sd_geometric' in content else None
+            geometric = _create_TCT_object(name, nm, filters, br_geom, sd_geom, calib, sun)
+            if 'spherical_albedo' in content:
+                where, how = content['spherical_albedo']
+                spherical = geometric.scaled_at(where, *aux.parse_value_sd(how))
+            elif 'bond_albedo' in content:
+                spherical = geometric.scaled_at(sun_filter, *aux.parse_value_sd(content['bond_albedo']))
+        if 'br_spherical' in content:
+            br_sphe = content['br_spherical']
+            sd_sphe = aux.higher_dim(content['sd_spherical'], len(br_sphe), axis=0) if 'sd_spherical' in content else None
+            spherical = _create_TCT_object(name, nm, filters, br_sphe, sd_sphe, calib, sun)
+            if 'geometric_albedo' in content:
+                where, how = content['geometric_albedo']
+                geometric = spherical.scaled_at(where, *aux.parse_value_sd(how))
+        if geometric is None and spherical is None:
+            print(f'# Note for the database object "{name}"')
+            print(f'- No brightness data. Spectrum stub object was created.')
+            TCT_obj = Spectrum.stub(name)
+    else:
+        TCT_obj = _create_TCT_object(name, nm, filters, br, sd, calib, sun)
+        # Non-specific albedo parsing
+        if 'albedo' in content:
+            if isinstance(content['albedo'], bool) and content['albedo']:
+                geometric = spherical = TCT_obj
+            elif isinstance(content['albedo'], Sequence):
+                where, how = content['albedo']
+                geometric = spherical = TCT_obj.scaled_at(where, *aux.parse_value_sd(how))
+            else:
+                print(f'# Note for the database object "{name}"')
+                print(f'- Invalid albedo value: {content["albedo"]}. Must be boolean or [filter/nm, [br, (sd)]].')
+        # Geometric albedo parsing
+        if 'geometric_albedo' in content:
+            if isinstance(content['geometric_albedo'], bool) and content['geometric_albedo']:
+                geometric = TCT_obj
+            elif isinstance(content['geometric_albedo'], Sequence):
+                where, how = content['geometric_albedo']
+                geometric = TCT_obj.scaled_at(where, *aux.parse_value_sd(how))
+            else:
+                print(f'# Note for the database object "{name}"')
+                print(f'- Invalid geometric albedo value: {content["geometric_albedo"]}. Must be boolean or [filter/nm, [br, (sd)]].')
+        # Spherical albedo parsing
+        if 'spherical_albedo' in content:
+            if isinstance(content['spherical_albedo'], bool) and content['spherical_albedo']:
+                spherical = TCT_obj
+            elif isinstance(content['spherical_albedo'], Sequence):
+                where, how = content['spherical_albedo']
+                spherical = TCT_obj.scaled_at(where, *aux.parse_value_sd(how))
+            else:
+                print(f'# Note for the database object "{name}"')
+                print(f'- Invalid spherical albedo value: {content["spherical_albedo"]}. Must be boolean or [filter/nm, [br, (sd)]].')
+        elif 'bond_albedo' in content:
+            spherical = TCT_obj.scaled_at(sun_filter, *aux.parse_value_sd(content['bond_albedo']))
+    tags = set()
+    if 'tags' in content:
+        for tag in content['tags']:
+            tags |= set(tag.split('/'))
+    if geometric or spherical:
+        phase_integral = phase_integral_sd = None
+        if 'phase_integral' in content:
+            phase_integral, phase_integral_sd = aux.parse_value_sd(content['phase_integral'])
+        elif 'phase_function' in content:
+            phase_func = content['phase_function']
+            if isinstance(phase_func, Sequence) and len(phase_func) == 2 and isinstance(phase_func[0], str) and isinstance(phase_func[1], dict):
+                phase_func, params = content['phase_function']
+                phase_integral, phase_integral_sd = aux.phase_function2phase_integral(phase_func, params)
+            else:
+                print(f'# Note for the database object "{name}"')
+                print(f'- Invalid phase function value: {content["phase_function"]}. Must be in the form ["name", {{param1: value1, ...}}]')
+        if phase_integral is not None:
+            phase_integral = (phase_integral, phase_integral_sd)
+        return ReflectiveBody(name, tags, geometric, spherical, phase_integral)
+    else:
+        return NonReflectiveBody(name, tags, TCT_obj)
+
+
+
+# ------------ Color Processing Section ------------
+# Needs improvement, see https://github.com/Askaniy/TrueColorTools/issues/22
+
+class ColorSystem:
+
+    def __init__(self, red: Sequence, green: Sequence, blue: Sequence, white: Sequence):
+        """
+        Initialise the ColorSystem object.
+        The implementation is based on https://scipython.com/blog/converting-a-spectrum-to-a-colour/
+        Defining the color system requires four 2d vectors (primary illuminants and the "white point")
+        """
+        self.red, self.green, self.blue, self.white = map(self.xy2xyz, [red, green, blue, white]) # chromaticities
+        self.M = np.vstack((self.red, self.green, self.blue)).T # the chromaticity matrix (rgb -> xyz) and its inverse
+        self.MI = np.linalg.inv(self.M) # white scaling array
+        self.wscale = self.MI.dot(self.white) # xyz -> rgb transformation matrix
+        self.T = self.MI / self.wscale[:, np.newaxis]
+    
+    @staticmethod
+    def xy2xyz(xy):
+        return np.array((xy[0], xy[1], 1-xy[0]-xy[0])) # (x, y, 1-x-y)
+
+# Used white points
+illuminant_E = (1/3, 1/3)
+#illuminant_D65 = (0.3127, 0.3291)
+
+# Used color systems
+srgb_system = ColorSystem((0.64, 0.33), (0.30, 0.60), (0.15, 0.06), illuminant_E)
+#hdtv_system = ColorSystem((0.67, 0.33), (0.21, 0.71), (0.15, 0.06), illuminant_D65)
+#smpte_system = ColorSystem((0.63, 0.34), (0.31, 0.595), (0.155, 0.070), illuminant_D65)
+
+# Stiles & Burch (1959) 2-deg color matching data, direct experimental data
+# http://www.cvrl.org/stilesburch2_ind.htm
+# Edge sensitivity modulo values less than 10⁴ were previously removed
+rgb_cmf = FilterSystem.from_list(('StilesBurch2deg.r', 'StilesBurch2deg.g', 'StilesBurch2deg.b'))
+
+# CIE XYZ functions transformed from the CIE (2006) LMS functions, 2-deg
+# http://www.cvrl.org/ciexyzpr.htm
+# Edge sensitivity values less than 10⁴ were previously removed
+xyz_cmf = FilterSystem.from_list(('cie2deg.x', 'cie2deg.y', 'cie2deg.z')) / 339.12
+# TODO: find a correct way to calibrate brightness for albedo!
+# 339.12 was guessed so that the equal-energy spectrum of unit brightness has color (1, 1, 1)
+
+# to be chosen depending on the bool flag (`cie`)
+cmfs = (rgb_cmf, xyz_cmf)
+
+
+class ColorImage:
+    """
+    A class for working with three-channel images.
+    Internal representation as a numpy array of the shape (3, hight, width) with values between 0 and 1.
+    (Note that the PhotospectralCube class has a transposed order of the axes.)
+    """
+
+    def __init__(self, br: Sequence, maximize_brightness=True):
+        """
+        The albedo flag on means that you have already normalized the brightness over the range.
+        By default, initialization implies normalization and you get chromaticity.
+
+        Args:
+        - `rgb` (Sequence): array of three values that are red, green and blue
+        - `maximize_brightness` (bool): to find the maximum value of the array and divide by it
+        """
+        self.br = np.clip(np.nan_to_num(br), 0, None, dtype='float')
+        if maximize_brightness and br.max() != 0:
+            self.br /= self.br.max()
+
+    @staticmethod
+    def from_spectral_data(spectral_cube: SpectralCube, maximize_brightness=True, srgb=False):
+        """ Convolves the spectral cube with one of the available CMF systems """
+        # TODO: add sRGB support for images!
+        return ColorImage((spectral_cube @ rgb_cmf).br, maximize_brightness)
+
+    def gamma_corrected(self):
+        """ Creates a new Color object with applied gamma correction """
+        other = deepcopy(self)
+        other.br = aux.gamma_correction(other.br)
+        return other
+    
+    def grayscale(self):
+        """ Converts rgb values to grayscale using sRGB luminance of the CIE 1931 """
+        return np.dot(self.br, (0.2126, 0.7152, 0.0722))
+
+
+class ColorPoint(ColorImage):
+    """
+    A class for working with three-channel point.
+    Internal representation as a numpy array of the shape (3) with values between 0 and 1.
+    """
+
+    @staticmethod
+    def from_spectral_data(spectrum: Spectrum, maximize_brightness=True, srgb=False):
+        """ Convolves the spectrum with one of the available CMF systems """
+        if srgb:
+            xyz = (spectrum @ xyz_cmf).br
+            rgb = srgb_system.T.dot(xyz)
+            if np.any(rgb < 0):
+                print(f'# Note for the Color object "{spectrum.name}"')
+                print(f'- RGB derived from XYZ turned out to be outside the color space: rgb={rgb}')
+                rgb -= rgb.min()
+                print(f'- Approximating by desaturating: rgb={rgb}')
+        else:
+            rgb = (spectrum @ rgb_cmf).br
+        return ColorPoint(rgb, maximize_brightness)
+
+    def to_bit(self, bit: int) -> np.ndarray:
+        """ Returns rounded color array, scaled to the appropriate power of two """
+        return self.br * (2**bit - 1)
+
+    def to_html(self):
+        """ Converts fractional rgb values to HTML-style hex string """
+        html = '#{:02x}{:02x}{:02x}'.format(*self.to_bit(8).round().astype('int'))
+        if len(html) != 7:
+            #print(f'# Note for the Color object "{self.name}"')
+            #print(f'- HTML-style color code feels wrong: {html}')
+            html = '#FFFFFF'
+            #print(f'- It has been replaced with {html}.')
+        return html
