+++ src/database.py
@@ -7,6 +7,7 @@ from typing import Sequence
 from json5 import load as json5load
 from pathlib import Path
 from traceback import format_exc
+
 from src.core import *
